+++ taskchampion/src/storage/mod.rs
@@ -29,7 +29,7 @@ pub use crate::server::VersionId;
 /// The default for base_version.
 pub(crate) const DEFAULT_BASE_VERSION: Uuid = crate::server::NO_VERSION_ID;
 
-/// A TaskStorage transaction, in which storage operations are performed.
+/// A Storage transaction, in which storage operations are performed.
 ///
 /// # Concurrency
 ///
@@ -40,9 +40,9 @@ pub(crate) const DEFAULT_BASE_VERSION: Uuid = crate::server::NO_VERSION_ID;
 /// # Commiting and Aborting
 ///
 /// A transaction is not visible to other readers until it is committed with
-/// [`crate::taskstorage::TaskStorageTxn::commit`].  Transactions are aborted if they are dropped.
+/// [`crate::storage::StorageTxn::commit`].  Transactions are aborted if they are dropped.
 /// It is safe and performant to drop transactions that did not modify any data without committing.
-pub trait TaskStorageTxn {
+pub trait StorageTxn {
     /// Get an (immutable) task, if it is in the storage
     fn get_task(&mut self, uuid: Uuid) -> Fallible<Option<TaskMap>>;
 
@@ -102,8 +102,8 @@ pub trait TaskStorageTxn {
 }
 
 /// A trait for objects able to act as task storage.  Most of the interesting behavior is in the
-/// [`crate::taskstorage::TaskStorageTxn`] trait.
-pub trait TaskStorage {
+/// [`crate::storage::StorageTxn`] trait.
+pub trait Storage {
     /// Begin a transaction
-    fn txn<'a>(&'a mut self) -> Fallible<Box<dyn TaskStorageTxn + 'a>>;
+    fn txn<'a>(&'a mut self) -> Fallible<Box<dyn StorageTxn + 'a>>;
 }
